requirements.txt
django>="4.2.2"
pillow>="10.0.0"
pillow-avif-plugin>="1.3.1"
django-autoslug>="1.9.9"
django-ckeditor>="6.7.0"